saathi-api/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
python-multipart
PyPDF2
python-docx
deep-translator
openpyxl
pytz
psutil
pyperclip
pygetwindow
pyttsx3
